Code Dump
Here's the version of GRQ I have, dated to March 2015.

Hint: I was 14 when I made this, it's not very well made!!!! However, it does still function and build (somehow!?) so it can't be that bad ;)

I may one day cme back and remake this but in Swift, with far more experience behind me.